+++ kcommons/src/main/java/com/kairos/dto/gdpr/metadata/AssetTypeBasicDTO.java
@@ -12,7 +12,7 @@ import java.util.List;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class AssetTypeBasicDTO {
 
-    private BigInteger id;
+    private Long id;
     @NotBlank(message = "error.message.name.notNull.orEmpty")
     @Pattern(message = "error.message.number.and.special.character.notAllowed", regexp = "^[a-zA-Z\\s]+$")
     private String name;
@@ -20,9 +20,9 @@ public class AssetTypeBasicDTO {
     private List<AssetTypeBasicDTO> subAssetTypes = new ArrayList<>();
 
 
-    public BigInteger getId() { return id; }
+    public Long getId() { return id; }
 
-    public void setId(BigInteger id) { this.id = id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name.trim(); }
